requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pkg-resources==0.0.0
PyYAML==3.13
six==1.11.0
Werkzeug==0.14.1
